Add files via upload
Removed distance and prominence from find_peaks call. This prevents the zero layers issue, although these parameters could be useful in certain situations. At the moment they are still a part of cloud_fuzz function, but these variables are not used at al. I am considering removing them altogether, but they may be useful later. 

Increasing the radius really improves resulting layer counts. I added a comment about this and increased the radius in the example.

I added an additional output scalar field: the total number of points found in the cylinder. this is useful for analysis, and will be useful to test the dot product cylinder search method